services/backend-server/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.4
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
psycopg2-binary==2.8.6
pytest==6.2.1
pytest-django==4.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.2.0
pytest-cov==2.10.1
black==20.8b1